add form mongoose schema. update file names for clarity.
