+++ client/components/form/Form.jsx
@@ -50,7 +50,9 @@ export default function Form() {
   const [lateTime, setLateTime] = useState(
     dayjs().hour(10).minute(0).second(0)
   );
-  const [address, setAddress] = useState([]);
+  //! Adding default fake address to pass validation only
+  //! Update this once the address validator is fully functional
+  const [address, setAddress] = useState("123 Main St, Los Angeles, CA 90012");
   // error state
   const [errorMsg, setErrorMsg] = useState("");
   const [errorPath, setErrorPath] = useState("");
