+++ server/src/routes/form.route.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { newAppt } from "../controllers/form.js";
+import { newAppt } from "../controllers/form.controller.js";
 
 const router = Router();
